uclibc: Enable wchar support to fix c++ string with stage2
When building libc++ it is configured according to what the libc
provide. However in stage2 the libc is configured differently, so if
libc++ provide some replacement type that are provided by the stage2
libc things break because of double definitions.

In the long run we'll need to either drop stage1 or have the libc to
be part of stage1 and allow stage2 to work with an extern libc.

Change-Id: I41c2b451161b2c7aa77c9f5968d5ae46c9756dce
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/41
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>