gcc: arm: Fix the softfp toolchain for SoC without FPU
When 'hardfp' toolchains where introduced the 'softfp' mode was
enabled for the non-hard toolchains. However the 'softfp' mode is just
using the FPU with a calling convention compatible with the 'soft'
mode. So it still require an FPU, on CPUs without any FPU the 'soft'
mode must be used instead.

Change-Id: I9e93ddb089c2285c4b0647478440133086fe9ca2
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1118
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>