support/gcc: don't overwrite LD_LIBRARY_PATH, add a prefix instead
The gcc wrapper script modifies LD_LIBRARY_PATH so that the linker uses
the libraries from ~/pbs-stage1/lib when running gcc.  Change that to
prefix the desired path to LD_LIBRARY_PATH instead of overwriting the
variable.

This fixes a lot of ugly warnings in stage2, where libfakeroot is not
found.

Change-Id: Ib1a28b2a5322c64169647683dd560ed45502b36f
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/2543
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>