packages/pkg-config: rename from pkgconfig
Since v0.18.1 the package is called pkg-config, with a dash. Using the
old name in the path still works as long as no patches are required.
Rename the package to avoid issues when there will be a patch needed in
the future. Keep the old Makefile target to ensure backwards
compatibility.

Change-Id: I7a3764662433fc1f5a053f4f0129f714dfd6cd5e
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6920
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>